Did a little spring cleaning on the two style sheets and readme.html that I updated and committed yesterday.
Forgot to add a log message on revision 23, so here it is. I updated the style.css. The default theme now looks nicer. Added a boarder around the images that are aggregated for the photos module. 

Mobile.css had some tweaks. Minor. Color and margin/padding.

Updated the readme.html. Now gives information on how to install Pez on a server, contact info, and credits!

These two revisions improve your social standing with the ladies. <3